Fold text_builder into renderer
Now that there is a 1-1 correlation between renderers and texture atlas's
we need to make it clear how to generate text that a renderer can process.
Making the text builder a part of the renderer seems to be the best way
since it should be intuitive that you should not pass font's generated
by one renderer into another.