Working C API with signed distance fields
The the bare minimum code to integrate GL and Freetype without
involving the user in all of the details. Went a super simple route by
having a single character per texture layer and a single large
128x128xN texture array. User just writes directly into the glyph
instances. Made this easier by applying the glyph top and left offset
in the shader.