Allow user to specify UV maps on a per material basis
Since with this change required that UV maps must be changed on the fly
the GL buffer management strategy has changed. The geometry structure
has been abandoned for now in favor of a mesh structure that contains
a buffer handle per UV map.

The b2l format itself needed to be changed as well to store separate
texcoord and tangent arrays per UV map layer. The chicken model has been
updated to the new format.