Add model viewer
Based on material editor code with Gtk/LGI interactions removed.
Depends on glib for g_strdup_printf()/g_free() and GdkPixbuf
but unlike the model viewer has no Gtk dependencies.

The model viewer accepts scene and object name on the command line.
An arbitrary scene and object will be chosen if none are specified.
The model will not animate but a frame # can be passed on the
command line and if the model is skinned the pose for that frame will
be selected.